+++ CHANGELOG.md
@@ -11,6 +11,7 @@ These changes are awaiting release:
 - Show `End Year` for TV Shows that have ended or been canceled.
 - Add full release date and last release date to their respective year elements as titles (browser tooltip).
 - Search: Support inline filters for Non-Multi searches (ex: You can now search `Spider Man year:2026` or `Spider Man fyear:2026` if you care strictly about the first release year).
+  - You'll find a new `filters` button on the search page that opens a modal explaining usage.
 
 ## Changed